requirements/base.txt
pytz==2019.1  # https://github.com/stub42/pytz
python-slugify==3.0.2  # https://github.com/un33k/python-slugify
Pillow==6.0.0  # https://github.com/python-pillow/Pillow
rcssmin==1.0.6 --install-option="--without-c-extensions"  # https://github.com/ndparker/rcssmin
argon2-cffi==19.1.0  # https://github.com/hynek/argon2_cffi
whitenoise==4.1.2  # https://github.com/evansd/whitenoise

# Django
# ------------------------------------------------------------------------------
django==2.0.13  # pyup: < 2.1  # https://www.djangoproject.com/
django-environ==0.4.5  # https://github.com/joke2k/django-environ
django-model-utils==3.1.2  # https://github.com/jazzband/django-model-utils
django-allauth==0.39.1  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.7.2  # https://github.com/django-crispy-forms/django-crispy-forms

# Django REST Framework
djangorestframework==3.9.2  # https://github.com/encode/django-rest-framework
coreapi==2.3.3  # https://github.com/core-api/python-client

django-tables2==2.0.6
django-taggit==1.1.0
django_tinymce-2.8.0
django-cache-memoize==0.1.6
django-admin-honeypot==1.0.0
taggit-selectize==2.6.0
django-treebeard==4.3
